cleaned up the interface module and restructured
