+++ src/qrisp/interface/converter/pennylane_converter.py
@@ -1,10 +1,26 @@
-from qrisp import QuantumCircuit, QuantumVariable, cx , x, rz
-from qrisp import ControlledOperation
-from qrisp.interface import BackendClient
-import numpy as np
+"""
+\********************************************************************************
+* Copyright (c) 2023 the Qrisp authors
+*
+* This program and the accompanying materials are made available under the
+* terms of the Eclipse Public License 2.0 which is available at
+* http://www.eclipse.org/legal/epl-2.0.
+*
+* This Source Code may also be made available under the following Secondary
+* Licenses when the conditions for such availability set forth in the Eclipse
+* Public License, v. 2.0 are satisfied: GNU General Public License, version 2
+* with the GNU Classpath Exception which is
+* available at https://www.gnu.org/software/classpath/license.html.
+*
+* SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+********************************************************************************/
+"""
+
 import types
-    
-import pennylane as qml 
+
+import numpy as np
+
+from qrisp.circuit import ControlledOperation
 
 """
 TODO: 
@@ -18,6 +34,7 @@ BUGS:
 
 
 def create_qml_instruction(op):
+    import pennylane as qml 
 
     #create basic, straight forward instructions
     if op.name == "rxx":
