+++ src/qrisp/interface/converter/pytket_converter.py
@@ -1,9 +1,24 @@
+"""
+\********************************************************************************
+* Copyright (c) 2023 the Qrisp authors
+*
+* This program and the accompanying materials are made available under the
+* terms of the Eclipse Public License 2.0 which is available at
+* http://www.eclipse.org/legal/epl-2.0.
+*
+* This Source Code may also be made available under the following Secondary
+* Licenses when the conditions for such availability set forth in the Eclipse
+* Public License, v. 2.0 are satisfied: GNU General Public License, version 2
+* with the GNU Classpath Exception which is
+* available at https://www.gnu.org/software/classpath/license.html.
+*
+* SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+********************************************************************************/
+"""
 
-from qrisp import QuantumCircuit
-from qrisp.core import QuantumVariable
-from qrisp import ControlledOperation
-from qrisp.interface import BackendClient
 import numpy as np
+
+from qrisp import ControlledOperation
 
 def create_tket_instruction(op):
     from pytket import OpType
