+++ src/qrisp/interface/converter/qulacs_converter.py
@@ -1,9 +1,28 @@
-from qulacs import QuantumState, QuantumCircuit
-import qulacs.gate as gt
+"""
+\********************************************************************************
+* Copyright (c) 2023 the Qrisp authors
+*
+* This program and the accompanying materials are made available under the
+* terms of the Eclipse Public License 2.0 which is available at
+* http://www.eclipse.org/legal/epl-2.0.
+*
+* This Source Code may also be made available under the following Secondary
+* Licenses when the conditions for such availability set forth in the Eclipse
+* Public License, v. 2.0 are satisfied: GNU General Public License, version 2
+* with the GNU Classpath Exception which is
+* available at https://www.gnu.org/software/classpath/license.html.
+*
+* SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+********************************************************************************/
+"""
 
 
 def qulacs_converter(qc):
+    
     from qrisp.simulator import extract_measurements
+
+    from qulacs import QuantumCircuit
+    import qulacs.gate as gt
 
     temp_qc, measurements = extract_measurements(qc.copy())
